rename tests stocksAPI and added db
